requirements.txt
about-time==4.2.1
aiobotocore==2.12.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.13
alive-progress==3.1.5
antlr4-python3-runtime==4.8
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
babel==2.14.0
backports-zoneinfo==0.2.1
biopython==1.79
blinker==1.7.0
botocore==1.34.51
botorch==0.8.5
chemprop==1.6.1
cachetools==5.0.0
cffi==1.16.0
chemprop==1.6.1
click==8.1.7
cloudpickle==1.2.2
cma==2.7.0
conda-pack==0.7.1
cython==3.0.0a10
deepchem==2.6.0.dev20211207173323
deprecated==1.2.14
descriptastorus==2.6.1
dill==0.3.7
django==4.2.9
docstring-parser==0.15
docutils==0.20.1
evoxbench==1.0.3
fasteners==0.19
filelock==3.13.1
fire==0.6.0
flask==3.0.0
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.3
glfw==2.7.0
gpytorch==1.10
grapheme==0.6.0
gym==0.14.0
higher==0.2.1
huggingface-hub==0.0.8
hydra-core==1.1.0.dev6
hyperopt==0.2.7
imageio==2.34.0
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.4.0
itsdangerous==2.1.2
jinja2==3.1.3
jmespath==1.0.1
linear-operator==0.4.0
markupsafe==2.1.3
mpmath==1.3.0
mujoco-py==2.1.2.14
multidict==6.0.5
multipledispatch==1.0.0
mypy-extensions==1.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.1.0.dev26
opt-einsum==3.3.0
packaging==23.2
pandas-flavor==0.6.0
pip==24.0
protobuf==4.25.2
py4j==0.10.9.7
pycparser==2.21
pyglet==1.3.2
pygments==2.17.2
pymoo==0.5.0
pyro-api==0.1.2
pyro-ppl==1.8.6
pyyaml==6.0.1
randomname==0.2.1
rdkit==2023.9.4
rdkit-pypi==2022.9.5
regex==2023.12.25
s3fs==2024.3.1
sacremoses==0.1.1
selfies==2.0.0
snowballstemmer==2.2.0
sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
sympy==1.12
tabulate==0.8.9
tensorboardx==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.2.0
tokenizers==0.10.3
# torchvision==0.11.1
tqdm==4.66.1
transformers==4.6.0
triton==2.1.0
typed-argument-parser==1.9.0
typing-inspect==0.9.0
urllib3==1.26.18
werkzeug==3.0.1
wrapt==1.16.0
xarray==2023.1.0
yarl==1.9.4
zipp==3.17.0
# scipy==1.10.1
# scikit-learn==0.21.3
-i https://pypi.tuna.tsinghua.edu.cn/simple